+++ node/clients/catalog.ts
@@ -1,4 +1,9 @@
-import { ExternalClient, InstanceOptions, IOContext } from '@vtex/api'
+import {
+  ExternalClient,
+  FORWARDED_HOST_HEADER,
+  InstanceOptions,
+  IOContext,
+} from '@vtex/api'
 
 export interface GetProductsAndSkuIdsReponse {
   items: number[]
@@ -14,28 +19,35 @@ const PAGE_SIZE = 100
 
 export class Catalog extends ExternalClient {
   constructor(protected context: IOContext, options?: InstanceOptions) {
-    super(
-      `http://${context.account}.vtexcommercestable.com.br`,
-      context,
-      {
-        ...(options ?? {}),
-        headers: {
-          ...(options?.headers ?? {}),
-          'Content-Type': 'application/json',
-          'VtexIdclientAutCookie': context.authToken,
-          'X-Vtex-Use-Https': 'true',
-        },
-      }
-    )
+    super(`http://${context.account}.vtexcommercestable.com.br`, context, {
+      ...(options ?? {}),
+      headers: {
+        ...(options?.headers ?? {}),
+        'Content-Type': 'application/json',
+        VtexIdclientAutCookie: context.authToken,
+        'X-Vtex-Use-Https': 'true',
+      },
+    })
   }
 
-  public getProductsIds (page: number, salesChannels?: string[]): Promise<GetProductsAndSkuIdsReponse> {
+  public getProductsIds(
+    page: number,
+    salesChannels?: string[]
+  ): Promise<GetProductsAndSkuIdsReponse> {
     return this.http.get('/api/catalog_system/pvt/products/GetProductsIds', {
       params: {
-        ...(salesChannels ? { SalesChannelId: salesChannels.join(',')} : {}),
+        ...(salesChannels ? { SalesChannelId: salesChannels.join(',') } : {}),
         Active: true,
         Page: page,
         pageSize: PAGE_SIZE,
+      },
+    })
+  }
+
+  public getSitemap(host: string, path = 'sitemap.xml') {
+    return this.http.get(path, {
+      headers: {
+        [FORWARDED_HOST_HEADER]: host,
       },
     })
   }