+++ node/middlewares/generateMiddlewares/prepare.ts
@@ -1,8 +1,24 @@
 import { CONFIG_BUCKET, GENERATION_CONFIG_FILE } from '../../utils'
 
 export async function prepare(ctx: EventContext, next: () => Promise<void>) {
-  const { body, vtex: { logger }, clients: { vbase } } = ctx
+  const {
+    body,
+    vtex: { logger },
+    clients: { vbase },
+    state: { isCrossBorder },
+  } = ctx
+
   const { generationId } = body
+
+  if (!isCrossBorder) {
+    logger.info({
+      message: 'Skipping generation for non-cross-border tenant',
+      payload: body,
+    })
+
+    return
+  }
+
   if (!generationId) {
     logger.error({ message: 'Missing generation id', payload: body })
     return