+++ node/middlewares/sitemap.ts
@@ -129,8 +129,40 @@ const sitemapBindingIndex = async (ctx: Context) => {
 
 export async function sitemap(ctx: Context, next: () => Promise<void>) {
   const {
-    state: { matchingBindings, bindingAddress, rootPath, settings },
+    state: { isCrossBorder },
   } = ctx
+
+  if (isCrossBorder) {
+    await legacySitemap(ctx)
+  } else {
+    await catalogSitemap(ctx)
+  }
+
+  next()
+}
+
+async function legacySitemap(ctx: Context) {
+  const {
+    state: {
+      isCrossBorder,
+      matchingBindings,
+      bindingAddress,
+      rootPath,
+      settings,
+    },
+    vtex: { logger },
+  } = ctx
+
+  logger.info({
+    message: 'Fetching legacy sitemap',
+    payload: {
+      isCrossBorder,
+      matchingBindings,
+      bindingAddress,
+      rootPath,
+      ignoreBindings: settings.ignoreBindings,
+    },
+  })
 
   const hasBindingIdentifier = rootPath || bindingAddress
   let $: any
@@ -158,5 +190,32 @@ export async function sitemap(ctx: Context, next: () => Promise<void>) {
   }
 
   ctx.body = $.xml()
-  next()
+}
+
+async function catalogSitemap(ctx: Context) {
+  const {
+    clients: { catalog },
+    headers: { 'x-forwarded-host': forwardedHost },
+    state: { isCrossBorder, forwardedPath },
+    vtex: { logger },
+  } = ctx
+
+  try {
+    logger.info({
+      message: 'Fetching catalog sitemap',
+      payload: {
+        forwardedHost,
+        forwardedPath,
+        isCrossBorder,
+      },
+    })
+
+    const sitemapData = await catalog.getSitemap(forwardedHost, forwardedPath)
+
+    ctx.body = sitemapData
+  } catch (error) {
+    logger.error(`Error fetching catalog sitemap: ${error}`)
+    ctx.status = 500
+    ctx.body = 'Internal Server Error'
+  }
 }